ptR1App/python-backend/requirements.txt
actionlib==1.14.3
angles==1.9.14
base-local-planner==1.17.3
blinker==1.9.0
bondpy==1.8.7
camera-calibration==1.17.0
camera-calibration-parsers==1.12.1
catkin==0.8.12
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
contourpy==1.3.0
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler==0.12.1
diagnostic-analysis==1.12.1
diagnostic-common-diagnostics==1.12.1
diagnostic-updater==1.12.1
dynamic-reconfigure==1.7.6
filelock==3.19.1
flask==3.1.2
fonttools==4.60.0
fsspec==2025.9.0
gazebo-plugins==2.9.3
gazebo-ros==2.9.3
gencpp==0.7.2
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.1
gennodejs==2.0.2
genpy==0.6.18
idna==3.10
image-geometry==1.16.2
importlib-metadata==8.7.0
importlib-resources==6.5.2
interactive-markers==1.12.2
itsdangerous==2.2.0
jinja2==3.1.6
joint-state-publisher==1.15.2
joint-state-publisher-gui==1.15.2
kiwisolver==1.4.7
laser-geometry==1.6.8
MarkupSafe==3.0.2
matplotlib==3.9.4
message-filters==1.17.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
polars==1.33.1
psutil==7.1.0
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-qt-binding==0.4.6
PyYAML==6.0.2
qt-dotgraph==0.4.5
qt-gui==0.4.5
qt-gui-cpp==0.4.5
qt-gui-py-common==0.4.5
requests==2.32.5
resource-retriever==1.12.10
rosapi==0.11.17
rosbag==1.17.4
rosboost-cfg==1.15.10
rosbridge-library==0.11.17
rosbridge-server==0.11.17
rosclean==1.15.10
roscreate==1.15.10
rosgraph==1.17.4
roslaunch==1.17.4
roslib==1.15.10
roslint==0.12.0
roslz4==1.17.4
rosmake==1.15.10
rosmaster==1.17.4
rosmsg==1.17.4
rosnode==1.17.4
rosparam==1.17.4
rospy==1.17.4
rosservice==1.17.4
rostest==1.17.4
rostopic==1.17.4
rosunit==1.15.10
roswtf==1.17.4
rqt-action==0.4.11
rqt-bag==0.5.3
rqt-bag-plugins==0.5.3
rqt-console==0.4.14
rqt-dep==0.4.14
rqt-graph==0.4.16
rqt-gui==0.5.5
rqt-gui-py==0.5.5
rqt-image-view==0.4.19
rqt-launch==0.4.10
rqt-logger-level==0.4.13
rqt-moveit==0.5.13
rqt-msg==0.4.12
rqt-nav-view==0.5.8
rqt-plot==0.4.16
rqt-pose-view==0.5.13
rqt-publisher==0.4.12
rqt-py-common==0.5.5
rqt-py-console==0.4.12
rqt-reconfigure==0.5.7
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.14
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.2
rqt-service-caller==0.4.12
rqt-shell==0.4.13
rqt-srv==0.4.11
rqt-tf-tree==0.6.5
rqt-top==0.4.11
rqt-topic==0.4.15
rqt-web==0.4.11
rviz==1.14.26
scipy==1.13.1
sensor-msgs==1.13.2
six==1.17.0
smach==2.5.3
smach-ros==2.5.3
smclib==1.8.7
sympy==1.13.1
tf==1.13.4
tf-conversions==1.13.4
tf2-geometry-msgs==0.7.10
tf2-kdl==0.7.10
tf2-py==0.7.10
tf2-ros==0.7.10
topic-tools==1.17.4
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
triton==3.2.0
typing-extensions==4.15.0
ultralytics==8.3.202
ultralytics-thop==2.0.17
urllib3==2.5.0
websockets==15.0.1
werkzeug==3.1.3
xacro==1.14.20
zipp==3.23.0
